requirements.txt
pandas==2.0.3
numpy==1.24.3
python-telegram-bot==20.1
requests==2.31.0